backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
websockets==13.0
pydantic==2.9.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-dotenv==1.0.1
python-multipart==0.0.12
# ── Auth ─────────────────────
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
